[refactor] modify components
- add `DataItem` and `RouletteItem`
